# openapi.yaml
openapi: 3.1.0
info:
  title: Vovremya Mini App Public API
  version: 1.1.0
  summary: HTTP API сервиса core для мини-приложения «Вовремя» в MAX
  description: |
    Единственный источник истины внешнего HTTP-контракта (владелец — Разработчик B).
    Транспорт: HTTPS через edge (Caddy) на том же origin, что и статика мини-приложения.
    Формат: JSON UTF-8, snake_case; даты — `YYYY-MM-DD` (локальная дата организации),
    моменты времени — RFC 3339 UTC. Ошибки — `application/problem+json` (RFC 9457).
    Аутентификация: `Authorization: Bearer <session token>`; токен выдаёт `POST /sessions`
    после серверной проверки подписи `initData` MAX.
    Идемпотентность создания: клиент передаёт `id` (UUID v4) создаваемого ресурса.
    Повтор с тем же `id` и тем же содержимым возвращает 200 и существующий ресурс,
    с другим содержимым — 409 `CONFLICT_ID_REUSED`.
servers:
  - url: http://localhost:8080/api/v1
    description: Локальный запуск через Docker Compose (APP_ENV=local)
  - url: https://{domain}/api/v1
    description: Опубликованный стенд (APP_ENV=prod)
    variables:
      domain:
        default: app.vovremya.ru
        description: Фактический домен задаётся в EDGE_SITE_ADDRESS при развёртывании
security:
  - bearerAuth: []
...
paths:
  /sessions:
    post:
      tags: [session]
      operationId: createSession
      summary: Создать сессию по данным запуска MAX (initData)
      security: []
      x-min-role: none
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SessionCreateRequest' }
      responses:
        default: { $ref: '#/components/responses/Unexpected' }
        '201':
          description: Сессия создана
          headers:
            X-Request-Id: { $ref: '#/components/headers/X-Request-Id' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Session' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/Unavailable' }
...
  /me:
    get:
      tags: [session]
      operationId: getMe
      summary: Текущий пользователь, его организации и состояние канала напоминаний
      x-min-role: authenticated
      responses:
        default: { $ref: '#/components/responses/Unexpected' }
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Me' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
    delete:
      tags: [session]
      operationId: deleteMe
      summary: Удалить аккаунт и все организации, где пользователь — владелец
      x-min-role: authenticated
      responses:
        default: { $ref: '#/components/responses/Unexpected' }
        '204': { description: Аккаунт и связанные данные удалены }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /organizations:
    post:
      tags: [organizations]
      operationId: createOrganization
      summary: Создать организацию (текущий пользователь становится владельцем)
      x-min-role: authenticated
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OrganizationCreate' }
      responses:
        default: { $ref: '#/components/responses/Unexpected' }
        '201':
          description: Создана
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Organization' }
        '200':
          description: Повтор идемпотентного запроса, организация уже создана
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Organization' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /organizations/{organizationId}:
    parameters:
      - $ref: '#/components/parameters/OrganizationId'
    get:
      tags: [organizations]
      operationId: getOrganization
      summary: Организация и сводка статусов документов
      x-min-role: viewer
      responses:
        default: { $ref: '#/components/responses/Unexpected' }
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Organization' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [organizations]
      operationId: updateOrganization
      summary: Изменить профиль организации (оптимистичная блокировка по version)
      x-min-role: editor
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OrganizationUpdate' }
      responses:
        default: { $ref: '#/components/responses/Unexpected' }
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Organization' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [organizations]
      operationId: deleteOrganization
      summary: Удалить организацию со всеми документами, участниками и напоминаниями
      x-min-role: owner
      responses:
        default: { $ref: '#/components/responses/Unexpected' }
        '204': { description: Удалена }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /organizations/{organizationId}/suggestions:
    parameters:
      - $ref: '#/components/parameters/OrganizationId'
    get:
      tags: [organizations]
      operationId: listSuggestions
      summary: Типовые документы по профилю, ещё не добавленные в реестр
      x-min-role: viewer
      responses:
        default: { $ref: '#/components/responses/Unexpected' }
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuggestionList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /organizations/{organizationId}/documents:
    parameters:
      - $ref: '#/components/parameters/OrganizationId'
    get:
      tags: [documents]
      operationId: listDocuments
      summary: Реестр документов, отсортированный по сроку (ближайшие первыми, бессрочные в конце)
      x-min-role: viewer
      parameters:
        - name: status
          in: query
          required: false
          schema: { $ref: '#/components/schemas/DeadlineStatus' }
        - name: q
          in: query
          required: false
          description: Поиск подстроки в названии без учёта регистра
          schema: { type: string, minLength: 1, maxLength: 100 }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
        - name: cursor
          in: query
          required: false
          schema: { type: string, maxLength: 200 }
      responses:
        default: { $ref: '#/components/responses/Unexpected' }
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DocumentPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [documents]
      operationId: createDocument
      summary: Добавить документ (идемпотентно по id)
      x-min-role: editor
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DocumentCreate' }
      responses:
        default: { $ref: '#/components/responses/Unexpected' }
        '201':
          description: Создан
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Document' }
        '200':
          description: Повтор идемпотентного запроса
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Document' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /documents/{documentId}:
    parameters:
      - $ref: '#/components/parameters/DocumentId'
    get:
      tags: [documents]
      operationId: getDocument
      summary: Карточка документа с историей периодов и шагами продления
      x-min-role: viewer
      responses:
        default: { $ref: '#/components/responses/Unexpected' }
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Document' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [documents]
      operationId: updateDocument
      summary: Изменить реквизиты, исправить даты текущего периода, отступы напоминаний
      x-min-role: editor
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DocumentUpdate' }
      responses:
        default: { $ref: '#/components/responses/Unexpected' }
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Document' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [documents]
      operationId: deleteDocument
      summary: Удалить документ вместе с периодами и напоминаниями
      x-min-role: editor
      responses:
        default: { $ref: '#/components/responses/Unexpected' }
        '204': { description: Удалён }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /organizations/{organizationId}/members:
    parameters:
      - $ref: '#/components/parameters/OrganizationId'
    get:
      tags: [members]
      operationId: listMembers
      summary: Участники организации
      x-min-role: viewer
      responses:
        default: { $ref: '#/components/responses/Unexpected' }
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MemberList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: 'vvs_<43 символа base64url>'
  headers:
    X-Request-Id:
      description: Идентификатор запроса (генерирует edge), повторяется в problem.request_id
      schema: { type: string, maxLength: 64 }
...
  parameters:
    OrganizationId:
      name: organizationId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    Unexpected:
      description: Иная ошибка (RATE_LIMITED, OVERLOADED, DEPENDENCY_UNAVAILABLE, INTERNAL)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
...
    SessionCreateRequest:
      type: object
      additionalProperties: false
      required: [init_data]
      properties:
        init_data:
          type: string
          minLength: 16
          maxLength: 4096
          description: Значение window.WebApp.initData без изменений
        platform: { $ref: '#/components/schemas/Platform' }
        app_version: { type: string, maxLength: 32 }
...